+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.1
-* added log viewer;
+* added the log viewer and the changelogs viewer;
 * added the changelog file.